Revert 'Fix SonarQube code quality issues - safe changes only'
This reverts commit 90f38e8014d19b680433482eb6dcf9c4088900de.

Reverting because the changes caused Quality Gate to fail.
Going back to the previous working state.

+++ backend/microservices/Productivity/src/main/java/com/freelancing/productivity/service/GoalService.java
@@ -78,7 +78,7 @@ public class GoalService {
         dto.setTargetDate(goal.getTargetDate());
         dto.setTotalTasks(total);
         dto.setDoneTasks(done);
-        dto.setCompletionRate(total == 0 ? 0.0 : done / (double) total);
+        dto.setCompletionRate(total == 0 ? 0.0 : (double) done / (double) total);
         dto.setWeeklyVelocity(done / weeks);
         return dto;
     }
